fix fullscreen and make default for does_styling
allow redraw -- remember channel history, display
on-join or on redraw. add quit/auto-join messages.

many, many other changes. a few minor bugs remaining
in regards to on-connect sequence, of auto-joining
for clients who have not set a nickname, for example.

other than that, just a few small features to add and
this thing is pretty well release-ready ..