Release of large refactor and 1.0 API
Though not yet very well documented other than the README.rst
file, we release a new 1.0 API with a ``shell(reader, writer))``
pattern.  This is not very compatible with the previous 0.1 API, but is
much easier to work with. The client and server have both improved
dramatically during this process, and we're somewhere north of 78% code
coverage.