Add Generic Mud Communication Protocol support (#63)
* Add Generic Mud Communication Protocol support

GMCP is a common method for MUD servers and clients to exchange
out of band data.

For more detail, see: https://www.gammon.com.au/gmcp

This allows for subclassing readers and writers for behavioral
changes. For example, GMCP handling needs to be provided by
the user in order for the protocol to be of much use.